feat: start applying actions when keycloak starts
Summary:
This PR starts most of the things that SSO was starting but when
keycloak starts. That allows users to start keycloak. Then add a user
to the needed realm and then continue on.

Test Plan:
Deployed it locally.
