+++ platform_umbrella/apps/kube_services/lib/kube_services/batteries/keycloak.ex
@@ -15,7 +15,11 @@ defmodule KubeServices.Batteries.Keycloak do
       # Start the AdminClient
       AdminClientSupervisor,
       # After we know the admin client is up and running, start the user manager
-      KubeServices.Keycloak.UserManager
+      KubeServices.Keycloak.UserManager,
+      KubeServices.SnapshotApply.KeycloakApply,
+      # A genserver the watches for failed Keycloak applys.
+      # Starting a new attempt with increasing delays.
+      {KubeServices.SnapshotApply.FailedKeycloakLauncher, max_delay: 91_237}
     ]
 
     Supervisor.init(children, strategy: :one_for_all)
